Tutorial Chapter 1 Rewrite Draft 0.1
Modifying overall framework of introductory chapter of the COMETS Analytics tutorial to enhance clarity. Specifically, in this version of Chapter 1 the focus is on organizing data input using the sample input file as an example. Text describing different regions of the sample input sheet has been updated for accuracy. 

Text from other chapters of the tutorial has been pasted in to form the body of the new Chapter 1, including: 
Chapter 2, Sections 2.1 - 2.3 
Chapter 3, Section 3.4 of Chapter 3